+++ packages/vite/src/node/server/pluginContainer.ts
@@ -32,6 +32,7 @@ SOFTWARE.
 import fs from 'fs'
 import { join, resolve } from 'path'
 import { performance } from 'perf_hooks'
+import { createRequire } from 'module'
 import type {
   EmittedFile,
   InputOptions,
@@ -56,8 +57,9 @@ import type { FSWatcher } from 'chokidar'
 import colors from 'picocolors'
 import type * as postcss from 'postcss'
 import type { Plugin } from '../plugin'
-import { cleanUrl, combineSourcemaps } from '../utils'
 import {
+  cleanUrl,
+  combineSourcemaps,
   createDebugger,
   ensureWatchedFile,
   generateCodeFrame,
@@ -150,8 +152,14 @@ export async function createPluginContainer(
 
   const watchFiles = new Set<string>()
 
+  // TODO: use import()
+  const _require = createRequire(import.meta.url)
+
   // get rollup version
-  const rollupPkgPath = resolve(require.resolve('rollup'), '../../package.json')
+  const rollupPkgPath = resolve(
+    _require.resolve('rollup'),
+    '../../package.json'
+  )
   const minimalContext: MinimalPluginContext = {
     meta: {
       rollupVersion: JSON.parse(fs.readFileSync(rollupPkgPath, 'utf-8'))