+++ playground/env/vite.config.js
@@ -1,5 +1,9 @@
 const { defineConfig } = require('vite')
 
 module.exports = defineConfig({
-  envPrefix: ['VITE_', 'CUSTOM_PREFIX_']
+  base: '/env/',
+  envPrefix: ['VITE_', 'CUSTOM_PREFIX_'],
+  build: {
+    outDir: 'dist/env'
+  }
 })