+++ playground/worker/__tests__/es/es-worker.spec.ts
@@ -66,7 +66,7 @@ describe.runIf(isBuild)('build', () => {
   test('inlined code generation', async () => {
     const assetsDir = path.resolve(testDir, 'dist/es/assets')
     const files = fs.readdirSync(assetsDir)
-    expect(files.length).toBe(25)
+    expect(files.length).toBe(27)
     const index = files.find((f) => f.includes('main-module'))
     const content = fs.readFileSync(path.resolve(assetsDir, index), 'utf-8')
     const worker = files.find((f) => f.includes('my-worker'))
@@ -113,6 +113,12 @@ test('emit chunk', async () => {
   )
   expect(await page.textContent('.emit-chunk-dynamic-import-worker')).toMatch(
     '"A string/es/"'
+  )
+})
+
+test('url query worker', async () => {
+  expect(await page.textContent('.simple-worker-url')).toMatch(
+    'Hello from simple worker!'
   )
 })